revert 4c304b7b74cb. Leave new pagesize defaults for key and ptree.
BDB uses a 4k pagesize on it's own. the existing key & ptree defaults set
by SKS are 2kB and 512B. Even moving to BDB 4KB page would be an improvement.
The only breakage this new default would cause is with an existing DB without
pagesize and ptree_pagesize in sksconf. We think this could be adequately
covered in documentation and release announcement. Adding these two to sksconf
prevents breakage.